+++ CHANGELOG.md
@@ -6,6 +6,15 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.1.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
  
 
+
+### 0.3.10 - 28.05.2024
+### Added
+- Confusion Matrix definition for python code
+
+### Changed
+- Compilation changes for MacOS
+- Starting proper API for ALMA algorithm class, first with constructor
+
 ### 0.3.9 - 2.02.2024
  
 ### Added